WIP
Add 1 water example

QoL

Try more fun things

Unify code between two water samplers.

WIP

Add common utilities

CR comments

Add common utilities

Add common utilities

Add common utilities

Restructure plotting code

Update keep_idxs keeping all states to be string
